# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Quizawa",
        "description": "Listing the routes",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/classrooms": {
            "get": {
                "tags": [
                    "Classrooms"
                ],
                "description": "List of classrooms",
                "parameters": [
                    {
                        "name": "api_token",
                        "in": "query",
                        "description": "User authentication",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Classrooms"
                ],
                "description": "Store a classroom",
                "parameters": [
                    {
                        "name": "api_token",
                        "in": "query",
                        "description": "User authentication",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "name of classroom",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "stored",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "parameters are missing",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/classrooms/{classroom}": {
            "get": {
                "tags": [
                    "Classrooms"
                ],
                "description": "Show one classroom",
                "parameters": [
                    {
                        "name": "api_token",
                        "in": "query",
                        "description": "User authentication",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "classroom",
                        "in": "path",
                        "description": "ID of the classroom to return",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "parameters are missing",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Classrooms"
                ],
                "description": "delete one classroom",
                "parameters": [
                    {
                        "name": "api_token",
                        "in": "query",
                        "description": "User authentication",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "classroom",
                        "in": "path",
                        "description": "ID of the classroom to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "successful without content",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "parameters are missing",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "not found",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "Classrooms"
                ],
                "description": "Update one classroom",
                "parameters": [
                    {
                        "name": "api_token",
                        "in": "query",
                        "description": "User authentication",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "classroom",
                        "in": "path",
                        "description": "ID of the classroom to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "name of classroom",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "parameters are missing",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "not found",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
...
    },
...
}